pyproject.toml
[tool.poetry]
name = "api-agendor"
version = "0.1.0"
description = ""
authors = ["thiagonoq <thiago_neves@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
fastapi = "^0.111.1"
pymongo = "^4.8.0"
motor = "^3.5.1"
aiofiles = "^24.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
